Bot_v3_LangGraph/requirements.txt
langchain==0.3.1
langchain_community==0.3.1
langchain_core==0.3.6
langchain_openai==0.2.1
langsmith==0.1.129
langgraph==0.2.28
aiogram==3.13.0
gspread==6.1.2
google-api-python-client==2.147.0
google-cloud-speech==2.27.0
oauth2client==4.1.3
google-auth==2.35.0
pytz==2022.7